Continuing Development
Deleted the Chatbot class.  Declared corresponding length values for our
Keyword and Response Arrays.  Renamed some variables.  Etc.

I am working on generalizing all of these methods, in order to make them
less specific to the Alice prototype.  Most of the functionality is in
place to do so, but I want the main() method to be a collection of
methods itself, with everything generalized in order to make it closer
to an actual game loop.  Also, I want to prioritize the game loop based
on an execution stack.  More to come later.